+++ ios/AppLovinMAXNativeAdViewManager.mm
@@ -17,17 +17,9 @@ RCT_EXPORT_MODULE(AppLovinMAXNativeAdView)
 RCT_EXPORT_VIEW_PROPERTY(adUnitId, NSString)
 RCT_EXPORT_VIEW_PROPERTY(placement, NSString)
 RCT_EXPORT_VIEW_PROPERTY(customData, NSString)
-RCT_EXPORT_VIEW_PROPERTY(extraParameters, NSDictionary)
-RCT_EXPORT_VIEW_PROPERTY(localExtraParameters, NSDictionary)
-
-// Asset views
-RCT_EXPORT_VIEW_PROPERTY(titleView, NSNumber)
-RCT_EXPORT_VIEW_PROPERTY(advertiserView, NSNumber)
-RCT_EXPORT_VIEW_PROPERTY(bodyView, NSNumber)
-RCT_EXPORT_VIEW_PROPERTY(callToActionView, NSNumber)
-RCT_EXPORT_VIEW_PROPERTY(iconView, NSNumber)
-RCT_EXPORT_VIEW_PROPERTY(optionsView, NSNumber)
-RCT_EXPORT_VIEW_PROPERTY(mediaView, NSNumber)
+RCT_EXPORT_VIEW_PROPERTY(extraParameters, NSArray)
+RCT_EXPORT_VIEW_PROPERTY(strLocalExtraParameters, NSArray)
+RCT_EXPORT_VIEW_PROPERTY(boolLocalExtraParameters, NSArray)
 
 // Callback
 RCT_EXPORT_VIEW_PROPERTY(onAdLoadedEvent, RCTDirectEventBlock)
@@ -39,6 +31,8 @@ RCT_EXPORT_VIEW_PROPERTY(onAdRevenuePaidEvent, RCTDirectEventBlock)
 {
     return YES;
 }
+
+#ifndef RCT_NEW_ARCH_ENABLED
 
 - (UIView *)view
 {
@@ -60,5 +54,67 @@ RCT_EXPORT_METHOD(loadAd:(nonnull NSNumber *)viewTag)
         [nativeAdView loadAd];
     }];
 }
+
+RCT_EXPORT_METHOD(updateAssetView:(nonnull NSNumber *)viewTag assetViewTag:(NSInteger)assetViewTag assetViewName:(NSString *)assetViewName)
+{
+    [self.bridge.uiManager addUIBlock:^(__unused RCTUIManager *uiManager, NSDictionary<NSNumber *, UIView *> *viewRegistry) {
+        
+        UIView *view = viewRegistry[viewTag];
+        if ( ![view isKindOfClass: [AppLovinMAXNativeAdView class]] )
+        {
+            [[AppLovinMAX shared] log: @"Cannot find AppLovinMAXNativeAdView with tag %@", viewTag];
+            return;
+        }
+        
+        AppLovinMAXNativeAdView *nativeAdView = (AppLovinMAXNativeAdView *) view;
+        
+        if ( [assetViewName isEqualToString: @"TitleView"] )
+        {
+            [nativeAdView setTitleView: @(assetViewTag)];
+        }
+        else if ( [assetViewName isEqualToString: @"AdvertiserView"] )
+        {
+            [nativeAdView setAdvertiserView: @(assetViewTag)];
+        }
+        else if ( [assetViewName isEqualToString: @"BodyView"] )
+        {
+            [nativeAdView setBodyView: @(assetViewTag)];
+        }
+        else if ( [assetViewName isEqualToString: @"CallToActionView"] )
+        {
+            [nativeAdView setCallToActionView: @(assetViewTag)];
+        }
+        else if ( [assetViewName isEqualToString: @"IconView"] )
+        {
+            [nativeAdView setIconView: @(assetViewTag)];
+        }
+        else if ( [assetViewName isEqualToString: @"OptionsView"] )
+        {
+            [nativeAdView setOptionsView: @(assetViewTag)];
+        }
+        else if ( [assetViewName isEqualToString: @"MediaView"] )
+        {
+            [nativeAdView setMediaView: @(assetViewTag)];
+        }
+    }];
+}
+
+RCT_EXPORT_METHOD(renderNativeAd:(nonnull NSNumber *)viewTag)
+{
+    [self.bridge.uiManager addUIBlock:^(__unused RCTUIManager *uiManager, NSDictionary<NSNumber *, UIView *> *viewRegistry) {
+
+        UIView *view = viewRegistry[viewTag];
+        if ( ![view isKindOfClass: [AppLovinMAXNativeAdView class]] )
+        {
+            [[AppLovinMAX shared] log: @"Cannot find AppLovinMAXNativeAdView with tag %@", viewTag];
+            return;
+        }
+        
+        AppLovinMAXNativeAdView *nativeAdView = (AppLovinMAXNativeAdView *) view;
+        [nativeAdView renderNativeAd];
+    }];
+}
+
+#endif // RCT_NEW_ARCH_ENABLED
 
 @end